requirements.txt
tweepy==4.14.0
google-generativeai==0.3.2
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
playwright==1.42.0 